python-script/requirements.txt
happybase==3.0.0
kafka-python==2.0.2